Applied physics to the sphere & materials to all objects. Created sphere script.
Added rigidbody to the sphere, applied ice property to all maze walls
and sphere to ensure neat and fast movement of the sphere. Applied plane
under maze so that sphere doesnt fall through the maze. Applied colours.
Added sphere script to reset sphere to spawn location when it falls off
the maze.